Check and report problematic log collection settings
We actually need to do this somewhere else where we can enable and
disable log collection based on the setting changing at runtime,
but this is a start.
