+++ input/collector.go
@@ -1,0 +1,151 @@
+package input
+
+import (
+	"fmt"
+	"os"
+	"runtime"
+	"strconv"
+	"strings"
+
+	"github.com/pganalyze/collector/config"
+	"github.com/pganalyze/collector/state"
+	"github.com/pganalyze/collector/util"
+	"github.com/shirou/gopsutil/host"
+	"github.com/shirou/gopsutil/process"
+)
+
+func getMemoryRssBytes() uint64 {
+	pid := os.Getpid()
+
+	p, err := process.NewProcess(int32(pid))
+	if err != nil {
+		return 0
+	}
+
+	mem, err := p.MemoryInfo()
+	if err != nil {
+		return 0
+	}
+
+	return mem.RSS
+}
+
+func getCollectorStats() state.CollectorStats {
+	var memStats runtime.MemStats
+	runtime.ReadMemStats(&memStats)
+
+	return state.CollectorStats{
+		GoVersion:                runtime.Version(),
+		ActiveGoroutines:         int32(runtime.NumGoroutine()),
+		CgoCalls:                 runtime.NumCgoCall(),
+		MemoryHeapAllocatedBytes: memStats.HeapAlloc,
+		MemoryHeapObjects:        memStats.HeapObjects,
+		MemorySystemBytes:        memStats.Sys,
+		MemoryRssBytes:           getMemoryRssBytes(),
+	}
+}
+
+func getCollectorPlatform(globalCollectionOpts state.CollectionOpts, logger *util.Logger) state.CollectorPlatform {
+	hostInfo, err := host.Info()
+	if err != nil {
+		if globalCollectionOpts.TestRun {
+			logger.PrintVerbose("Could not get collector host information: %s", err)
+		}
+		return state.CollectorPlatform{}
+	}
+
+	var virtSystem string
+	if hostInfo.VirtualizationRole == "guest" {
+		virtSystem = hostInfo.VirtualizationSystem
+	}
+	return state.CollectorPlatform{
+		StartedAt:            globalCollectionOpts.StartedAt,
+		Architecture:         runtime.GOARCH,
+		Hostname:             hostInfo.Hostname,
+		OperatingSystem:      hostInfo.OS,
+		Platform:             hostInfo.Platform,
+		PlatformFamily:       hostInfo.PlatformFamily,
+		PlatformVersion:      hostInfo.PlatformVersion,
+		KernelVersion:        hostInfo.KernelVersion,
+		VirtualizationSystem: virtSystem,
+	}
+}
+
+func getCollectorConfig(c config.ServerConfig) state.CollectorConfig {
+	return state.CollectorConfig{
+		SectionName:             c.SectionName,
+		DisableLogs:             c.DisableLogs,
+		DisableActivity:         c.DisableActivity,
+		EnableLogExplain:        c.EnableLogExplain,
+		DbName:                  c.DbName,
+		DbUsername:              c.DbUsername,
+		DbHost:                  c.DbHost,
+		DbPort:                  int32(c.DbPort),
+		DbSslmode:               c.DbSslMode,
+		DbHasSslrootcert:        c.DbSslRootCert != "" || c.DbSslRootCertContents != "",
+		DbHasSslcert:            c.DbSslCert != "" || c.DbSslCertContents != "",
+		DbHasSslkey:             c.DbSslKey != "" || c.DbSslKeyContents != "",
+		DbExtraNames:            c.DbExtraNames,
+		DbAllNames:              c.DbAllNames,
+		AwsRegion:               c.AwsRegion,
+		AwsDbInstanceId:         c.AwsDbInstanceID,
+		AwsHasAccessKeyId:       c.AwsAccessKeyID != "",
+		AzureDbServerName:       c.AzureDbServerName,
+		AzureEventhubNamespace:  c.AzureEventhubNamespace,
+		AzureEventhubName:       c.AzureEventhubName,
+		AzureAdTenantId:         c.AzureADTenantID,
+		AzureAdClientId:         c.AzureADClientID,
+		AzureHasAdCertificate:   c.AzureADCertificatePath != "",
+		GcpCloudsqlInstanceId:   c.GcpCloudSQLInstanceID,
+		GcpPubsubSubscription:   c.GcpPubsubSubscription,
+		GcpHasCredentialsFile:   c.GcpCredentialsFile != "",
+		GcpProjectId:            c.GcpProjectID,
+		ApiSystemId:             c.SystemID,
+		ApiSystemType:           c.SystemType,
+		ApiSystemScope:          c.SystemScope,
+		DbLogLocation:           c.LogLocation,
+		DbLogDockerTail:         c.LogDockerTail,
+		IgnoreTablePattern:      c.IgnoreTablePattern,
+		IgnoreSchemaRegexp:      c.IgnoreSchemaRegexp,
+		QueryStatsInterval:      int32(c.QueryStatsInterval),
+		MaxCollectorConnections: int32(c.MaxCollectorConnections),
+		FilterLogSecret:         c.FilterLogSecret,
+		FilterQuerySample:       c.FilterQuerySample,
+		HasProxy:                c.HTTPProxy != "" || c.HTTPSProxy != "",
+		ConfigFromEnv:           os.Getenv("PGA_API_KEY") != "",
+	}
+}
+
+const MinSupportedLogMinDurationStatement = 100
+
+func getCollectorLogSnapshotStatus(c config.ServerConfig, globalCollectionOpts state.CollectionOpts, settings []state.PostgresSetting) (disabled bool, disabledReason string) {
+	if c.DisableLogs {
+		return true, "the collector setting disable_logs or environment variable PGA_DISABLE_LOGS is set"
+	}
+	var reasons []string
+	for _, setting := range settings {
+		if setting.Name == "log_min_duration_statement" && setting.CurrentValue.Valid {
+			numVal, err := strconv.Atoi(setting.CurrentValue.String)
+			if err != nil {
+				continue
+			}
+			if numVal < MinSupportedLogMinDurationStatement {
+				disabled = true
+				reason := fmt.Sprintf("log_min_duration_statement is set to '%d', below minimum supported threshold '%d'", numVal, MinSupportedLogMinDurationStatement)
+				reasons = append(reasons, reason)
+			}
+		} else if setting.Name == "log_duration" && setting.CurrentValue.Valid {
+			if setting.CurrentValue.String == "on" {
+				disabled = true
+				reasons = append(reasons, "log_duration is set to unsupported value 'on'")
+			}
+		} else if setting.Name == "log_statement" && setting.CurrentValue.Valid {
+			if setting.CurrentValue.String == "all" {
+				disabled = true
+				reasons = append(reasons, "log_statement is set to unsupported value 'all'")
+			}
+		}
+	}
+
+	return disabled, strings.Join(reasons, "; ")
+}
