+++ state/state.go
@@ -65,6 +65,9 @@ type TransientState struct {
 
 	CollectorConfig   CollectorConfig
 	CollectorPlatform CollectorPlatform
+
+	CollectorLogSnapshotDisabled       bool
+	CollectorLogSnapshotDisabledReason string
 }
 
 type CollectorConfig struct {
